Fix for missing temperature updates

+++ src/backend/Home.Core/Devices/IPressureSensor.cs
@@ -2,7 +2,7 @@ namespace Home.Core.Devices {
 
     public interface IPressureSensor {
 
-        int Pressure { get; }
+        double Pressure { get; }
 
     }
